+++ drivers/crypto/caam/caamrng.c
@@ -292,7 +292,7 @@ int caam_rng_init(struct device *ctrldev)
 	ctx->rng.init    = caam_init;
 	ctx->rng.cleanup = caam_cleanup;
 	ctx->rng.read    = caam_read;
-	ctx->rng.priv    = (unsigned long)ctx;
+	ctx->rng.priv    = (uintptr_t)ctx;
 
 	dev_info(ctrldev, "registering rng-caam\n");
