+++ drivers/gpu/drm/amd/amdkfd/kfd_device.c
@@ -1246,7 +1246,7 @@ static inline uint32_t *kfd_gtt_sa_calc_cpu_addr(void *start_addr,
 						unsigned int bit_num,
 						unsigned int chunk_size)
 {
-	return (uint32_t *) ((uint64_t) start_addr + bit_num * chunk_size);
+	return (uint32_t *) ((uintptr_t) start_addr + bit_num * chunk_size);
 }
 
 int kfd_gtt_sa_allocate(struct kfd_node *node, unsigned int size,
