+++ drivers/net/ethernet/mediatek/mtk_eth_soc.c
@@ -4864,7 +4864,7 @@ static int mtk_add_mac(struct mtk_eth *eth, struct device_node *np)
 	SET_NETDEV_DEV(eth->netdev[id], eth->dev);
 	eth->netdev[id]->watchdog_timeo = 5 * HZ;
 	eth->netdev[id]->netdev_ops = &mtk_netdev_ops;
-	eth->netdev[id]->base_addr = (unsigned long)eth->base;
+	eth->netdev[id]->base_addr = (uintptr_t)eth->base;
 
 	eth->netdev[id]->hw_features = eth->soc->hw_features;
 	if (eth->hwlro)
