+++ drivers/net/ethernet/meta/fbnic/fbnic_fw.c
@@ -160,7 +160,7 @@ static void fbnic_mbx_unmap_and_free_msg(struct fbnic_dev *fbd, int mbx_idx,
 	dma_unmap_single(fbd->dev, mbx->buf_info[desc_idx].addr,
 			 PAGE_SIZE, direction);
 
-	free_page((unsigned long)mbx->buf_info[desc_idx].msg);
+	free_page((uintptr_t)mbx->buf_info[desc_idx].msg);
 	mbx->buf_info[desc_idx].msg = NULL;
 }
 
@@ -212,7 +212,7 @@ static int fbnic_mbx_alloc_rx_msgs(struct fbnic_dev *fbd)
 		err = fbnic_mbx_map_msg(fbd, FBNIC_IPC_MBX_RX_IDX, msg,
 					FBNIC_RX_PAGE_SIZE, 0);
 		if (err)
-			free_page((unsigned long)msg);
+			free_page((uintptr_t)msg);
 	}
 
 	return err;
@@ -398,7 +398,7 @@ static int fbnic_fw_xmit_simple_msg(struct fbnic_dev *fbd, u32 msg_type)
 
 	err = fbnic_mbx_map_tlv_msg(fbd, msg);
 	if (err)
-		free_page((unsigned long)msg);
+		free_page((uintptr_t)msg);
 
 	return err;
 }
@@ -488,7 +488,7 @@ int fbnic_fw_xmit_ownership_msg(struct fbnic_dev *fbd, bool take_ownership)
 	return err;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -695,7 +695,7 @@ static int fbnic_fw_xmit_heartbeat_message(struct fbnic_dev *fbd)
 	return err;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -796,7 +796,7 @@ int fbnic_fw_xmit_fw_start_upgrade(struct fbnic_dev *fbd,
 	return 0;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -872,7 +872,7 @@ int fbnic_fw_xmit_fw_write_chunk(struct fbnic_dev *fbd,
 	return 0;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -975,7 +975,7 @@ int fbnic_fw_xmit_tsene_read_msg(struct fbnic_dev *fbd,
 	return 0;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -1093,7 +1093,7 @@ static void fbnic_mbx_process_rx_msgs(struct fbnic_dev *fbd)
 		dev_dbg(fbd->dev, "Parsed msg type %d\n", msg->hdr.type);
 next_page:
 
-		free_page((unsigned long)rx_mbx->buf_info[head].msg);
+		free_page((uintptr_t)rx_mbx->buf_info[head].msg);
 		rx_mbx->buf_info[head].msg = NULL;
 
 		head++;
